# [HttpFoundation] Add `StreamedJsonResponse` for efficient JSON streaming