@alexander-schranz 2022-09-27
The error in the tests of `Stopwatch` is unrelated to the pull request.