@ro0NL 2022-09-29
would it be reasonable to consider a "compute json inline" approach, rather than end-users taking care of unique identifiers

```php
$lazyJson = ['key' => fn() => yield from $heavy];
```