@stof 2022-09-29
@ro0NL this would force to re-implement the whole json encoding in userland