@ro0NL 2022-09-29
we could array walk the structure first, thus keeping the unique placeholders an implementation detail.