@stof 2022-09-29
@ro0NL if you do that, you are not streaming json anymore, defeating the whole purpose of this PR.