@ro0NL 2022-09-29
the idea is split the generators from the structure, preserving remaining logic. But this is an extra step yes, thus less ideal perhaps.