@alexander-schranz 2022-09-29
@ro0NL interesting input. As I think the structure array is mostly small it could be possible. But we would need to have a look at what difference this would be in the performance.

I hacked something together using `array_walk_recursive`: https://3v4l.org/tndhO. Will have a deeper look at it at the evening or next days.