@stof 2022-09-29
@alexander-schranz be careful when implementing this. `is_callable` would turn some strings into placeholders instead of outputting them.