@alexander-schranz 2022-09-29
@stof great hint think `$item instanceof Closure` should then do the job?