@stof 2022-09-29
now that we have first class callables, I would say yes. You can convert any callable to a closure using this feature.