@alexander-schranz 2022-09-29
Okay I don't need to check for `closures` or callables. I just need to check on `\Generators` because the Closures are already called. Which is very important, as example when Database connection is not available the exception need to be thrown in the Controller and should not be thrown when Status Code 200 is already returned:

```php
return new StreamedJsonResponse(
    [
        '_embedded' => [
            'articles' => $this->findArticles('Article'), // returns a \Generator which will generate a list of data
        ],
    ],
);
```

The diff between old and new implementation is not big it just takes about `0.0000128s` todo the array_walk_recursive and replace it. It also did not have any visible changes on the memory usage. The tested arrays are really small but that will mostly be the case I think in this kind of responses. 

I also update the example repository using the new class under `/symfony-articles.json`: [https://github.com/alexander-schranz/efficient-json-streaming-with-symfony-doctrine ](https://github.com/alexander-schranz/efficient-json-streaming-with-symfony-doctrine) if somebody want to experiment with it.