@alexander-schranz 2022-10-24
@nicolas-grekas I would keep a default flush, mostly the Generator will come from a Repository or other service and for a good DX we should not force them to create then additional function to wrap that generator just to add `flush` to it.