@nicolas-grekas 2022-10-24
Wrapping a generator is quite easy. How does that play with ob-handlers for example? What if flushing should happen every 4k bytes instead of every n-th items? etc.

Why do we need any flush control at all? We don't anywhere else and StreamedResponse even mentions calling flush() is userland's responsibility.