@alexander-schranz 2022-10-24
@nicolas-grekas I would prefer to have some default flush handling implemented and make it possible to overwrite it. But if you think we should not do that I'm fine to adjust the code, just think a default flush handling would be good if we provide such a class in the framework.