@welcoMattic 2022-10-24
I think we should remove flush control for now. If needed, we could add it later, and take time to do it properly. Ok for you @alexander-schranz, @nicolas-grekas?