@alexander-schranz 2022-10-24
@welcoMattic okay will adjust the PR at evening. Added a `TODO` for the docs PR so the docs will show a correct example and usage with flush for it.