@alexander-schranz 2022-10-24
@nicolas-grekas @jderusse review comments fixed. Will update the prototype for testing asap.