@OskarStark 2022-10-25
I propose to add the content from the README of your prototype application to the PR header 👍🏻