@dunglas 2022-11-25
For the record, @mtarld @soyuka and I are working on a new component that will be an alternative to `json_encode/json_decode` and to the Symfony Serializer that will natively support JSON streaming (for encoding and decoding). Maybe will it be possible to use this component in this PR.