@alexander-schranz 2022-12-29
🎉 Thx you all for the great feedback and ideas. Think we got a great solution out of it with a better DX as I could think of when created the Pull request.

@chalasr that sounds great :)